# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "https://github.com/junun/admin-go",
        "contact": {
            "name": "Junun",
            "url": "https://github.com/junun/admin-go",
            "email": "junun717@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/menus": {
            "get": {
                "description": "一级菜单列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "一级菜单列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "一级菜单添加， type must 1",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "一级菜单添加",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "一级菜单信息",
                        "name": "Data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.MenuResource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/menus/{id}": {
            "put": {
                "description": "一级菜单修改",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "一级菜单修改",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "一级菜单信息",
                        "name": "Data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.MenuResource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "菜单删除",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "菜单删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/admin/submenus": {
            "get": {
                "description": "二级菜单列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "二级菜单列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "二级菜单添加， type must 1",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "二级菜单添加",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "二级菜单信息",
                        "name": "Data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.SubMenuResource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/submenus/{id}": {
            "put": {
                "description": "二级菜单修改",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "二级菜单修改",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "二级菜单信息",
                        "name": "Data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.SubMenuResource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "code\": 500, \"message\": \"\", \"data\": {}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.MenuResource": {
            "type": "object",
            "properties": {
                "icon": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
...
        "admin.SubMenuResource": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    }
}